# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/customers": {
      "get": {
        "operationId": "CustomersController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "example": "João",
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PF",
                "PJ"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "not_requested",
                "requested",
                "in_review",
                "approved",
                "rejected",
                "suspended"
              ]
            }
          },
          {
            "name": "cpf",
            "required": false,
            "in": "query",
            "schema": {
              "example": "12345678901",
              "type": "string"
            }
          },
          {
            "name": "cnpj",
            "required": false,
            "in": "query",
            "schema": {
              "example": "12345678000190",
              "type": "string"
            }
          },
          {
            "name": "email",
            "required": false,
            "in": "query",
            "schema": {
              "example": "joao@email.com",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar customers",
        "tags": [
          "Customers"
        ]
      },
      "post": {
        "operationId": "CustomersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerLocalDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Criar customer",
        "tags": [
          "Customers"
        ]
      }
    },
    "/customers/me": {
      "get": {
        "operationId": "CustomersController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ver meu customer",
        "tags": [
          "Customers"
        ]
      }
    },
    "/customers/{id}": {
      "get": {
        "operationId": "CustomersController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ver customer por ID",
        "tags": [
          "Customers"
        ]
      },
      "patch": {
        "operationId": "CustomersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerLocalDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar customer",
        "tags": [
          "Customers"
        ]
      }
    },
    "/customers/{id}/balance": {
      "get": {
        "operationId": "CustomersController_balance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Saldo Pix do customer (soma dos depósitos Pix)",
        "tags": [
          "Customers"
        ]
      }
    },
    "/customers/{id}/statement": {
      "get": {
        "operationId": "CustomersController_statement",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "Data inicial (ISO 8601)",
            "schema": {
              "example": "2025-01-01T00:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "Data final (ISO 8601)",
            "schema": {
              "example": "2025-01-31T23:59:59.999Z",
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Alias para \"from\" (compatibilidade)",
            "schema": {
              "example": "2025-01-01",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "Alias para \"to\" (compatibilidade)",
            "schema": {
              "example": "2025-01-31",
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Tipo de transação",
            "schema": {
              "example": "PIX_IN",
              "type": "string",
              "enum": [
                "DEPOSIT",
                "WITHDRAW",
                "TRANSFER_IN",
                "TRANSFER_OUT",
                "PIX_IN",
                "PIX_OUT",
                "FEE",
                "REVERSAL"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Status da transação",
            "schema": {
              "example": "COMPLETED",
              "type": "string",
              "enum": [
                "PENDING",
                "PROCESSING",
                "COMPLETED",
                "FAILED",
                "REVERSED",
                "CANCELED"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Número da página",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Limite de resultados por página",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "example": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Extrato do customer",
        "tags": [
          "Customers"
        ]
      }
    },
    "/customers/{id}/kyc/request": {
      "post": {
        "operationId": "CustomersController_requestKyc",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Solicitar KYC - retorna URL de verificação Didit",
        "tags": [
          "Customers"
        ]
      }
    },
    "/customers/{id}/kyc/status": {
      "get": {
        "operationId": "CustomersController_kycStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Consultar status do KYC e decisão Didit",
        "tags": [
          "Customers"
        ]
      }
    },
    "/customers/{id}/kyc/review": {
      "patch": {
        "operationId": "CustomersController_review",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mover KYC para revisão",
        "tags": [
          "Customers"
        ]
      }
    },
    "/customers/{id}/kyc/approve": {
      "patch": {
        "operationId": "CustomersController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Aprovar KYC",
        "tags": [
          "Customers"
        ]
      }
    },
    "/customers/{id}/kyc/reject": {
      "patch": {
        "operationId": "CustomersController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Rejeitar KYC",
        "tags": [
          "Customers"
        ]
      }
    },
...
    "/didit/webhooks/verification": {
      "post": {
        "operationId": "DiditController_handleVerificationWebhook",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiditWebhookPayloadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook processado com sucesso"
          }
        },
        "summary": "Webhook para receber notificações de verificação Didit",
        "tags": [
          "Didit Webhooks"
        ]
      }
    },
...
    "/fdbank/customers/{id}": {
      "get": {
        "operationId": "FdbankCustomerController_getCustomer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "FdbankCustomer"
        ]
      },
      "put": {
        "operationId": "FdbankCustomerController_updateCustomer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "FdbankCustomer"
        ]
      },
      "delete": {
        "operationId": "FdbankCustomerController_deleteCustomer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "FdbankCustomer"
        ]
      }
    },
...
    "/pix-keys/customer/{customerId}": {
      "get": {
        "operationId": "PixKeysController_getKeysByCustomer",
        "parameters": [
          {
            "name": "customerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PixKeys"
        ]
      }
    },
...
  },
  "info": {
    "title": "OTSEM API",
    "description": "Documentação completa das rotas da OTSEM API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AddressDto": {
        "type": "object",
        "properties": {}
      },
      "CreateCustomerLocalDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "PF",
              "PJ"
            ],
            "example": "PF"
          },
          "name": {
            "type": "string",
            "example": "João Silva"
          },
          "email": {
            "type": "string",
            "example": "joao@email.com"
          },
          "phone": {
            "type": "string",
            "example": "+5511999999999"
          },
          "cpf": {
            "type": "string",
            "example": "12345678901"
          },
          "birthday": {
            "type": "string",
            "example": "1990-01-15"
          },
          "rg": {
            "type": "string",
            "example": "123456789"
          },
          "mothersName": {
            "type": "string",
            "example": "João Silva"
          },
          "cnpj": {
            "type": "string",
            "example": "12345678000190"
          },
          "companyName": {
            "type": "string",
            "example": "Empresa LTDA"
          },
          "tradingName": {
            "type": "string",
            "example": "Empresa"
          },
          "foundingDate": {
            "type": "string",
            "example": "2020-01-01"
          },
          "address": {
            "$ref": "#/components/schemas/AddressDto"
          },
          "metadata": {
            "type": "object",
            "example": {
              "referralCode": "XYZ123"
            }
          }
        },
        "required": [
          "type",
          "name",
          "email"
        ]
      },
      "UpdateCustomerLocalDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "João Silva"
          },
          "email": {
            "type": "string",
            "example": "joao@email.com"
          },
          "phone": {
            "type": "string",
            "example": "+5511999999999"
          },
          "birthday": {
            "type": "string",
            "example": "1990-01-15"
          },
          "rg": {
            "type": "string",
            "example": "123456789"
          },
          "mothersName": {
            "type": "string",
            "example": "João Silva"
          },
          "companyName": {
            "type": "string",
            "example": "Empresa LTDA"
          },
          "tradingName": {
            "type": "string",
            "example": "Empresa"
          },
          "foundingDate": {
            "type": "string",
            "example": "2020-01-01"
          },
          "address": {
            "$ref": "#/components/schemas/AddressDto"
          },
          "metadata": {
            "type": "object",
            "example": {
              "referralCode": "XYZ123"
            }
          },
          "accountStatus": {
            "type": "string",
            "enum": [
              "not_requested",
              "requested",
              "in_review",
              "approved",
              "rejected",
              "suspended"
            ]
          }
        }
      },
      "DiditWebhookPayloadDto": {
        "type": "object",
        "properties": {
          "session_id": {
            "type": "string",
            "description": "ID da sessão de verificação"
          },
          "status": {
            "type": "string",
            "description": "Status da sessão"
          },
          "decision": {
            "type": "string",
            "description": "Decisão da verificação: Approved, Declined, Review"
          },
          "vendor_data": {
            "type": "string",
            "description": "Dados do vendor passados na criação"
          },
          "document_data": {
            "type": "object"
          },
          "face_data": {
            "type": "object"
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "status"
        ]
      },
...
      "UpdateCustomerDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}